fix: moving interfaces to interfaces/form.ts and updating imports
